requirements/requirements.txt
Django==2.1.7
psycopg2==2.7.7
django-environ==0.4.5
django-redis==4.10.0
djangorestframework==3.9.2
django-allauth==0.39.1
django-filter==2.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
easy-thumbnails==2.6
django-taggit==1.1.0
django-taggit-serializer==0.1.7